Fix empty OpenAPI schemas which inherit from templated base types
This change fixes an issue where the OpenAPI emitter will generate an "empty"
schema definition for a model type that inherits from a single templated type
without adding any new properties of its own.  This "facade" definition causes
issues with ARM Swagger linters (and also looks pretty bad in the output
Swagger) so we're special-casing a fix to handle this narrow scenario.

Model types that extend multiple base types or non-templated base types are not
affected and will retain the previous behavior.